requirements.txt
pyz3r
ruamel.yaml
tzdata
disnake
python-dotenv
racetime-bot
apscheduler
tenacity
